timeconv/ add following features  - converting Float to Integer  - display usage  - test mode (-t);

+++ utils/timeconv.sh
@@ -1,12 +1,20 @@
 #!/bin/bash
 # Converting time field in Sqlog or Logfile (x 1000/Integer)
+[ "$1" ] || {
+    echo "Usage: timeconv (-t) [file(*.sq3,log) ...]"
+    exit;
+}
+if [ "$1" = '-t' ]; then
+    shift
+    _test=1
+fi
 tempdir=~/.var/temp
 [ -d $tempdir ] || mkdir $tempdir
 n='[0-9]'
 int="$n$n$n$n$n$n$n$n$n$n"
 dec="$n$n$n"
 conv(){
-    sed -e "s/['\"]*\($int\)\.*\($dec\)$n*['\"]*/\1\2/"
+    sed -e "s/['\"]*\($int\)\.*\($dec\)\.*$n*['\"]*/\1\2/"
 }
 
 for file ;do
@@ -14,14 +22,24 @@ for file ;do
     case $file in
         *.sq3)
             for i in `echo .tables|sqlite3 $file`; do
+                echo "   $i"
                 echo ".dump $i"|sqlite3 $file| conv > $tempdir/$i
-                (echo "drop table $i;";cat $tempdir/$i)|sqlite3 $file
+                if [ "$_test" ] ; then
+                    cat $tempdir/$i
+                else
+                    (echo "drop table $i;";cat $tempdir/$i)|sqlite3 $file
+                fi
                 rm $tempdir/$i
             done
             ;;
         *.log)
             <$file conv > $tempdir/$file
-            mv $tempdir/$file .
+            if [ "$_test" ] ; then
+                cat $tempdir/$file
+                rm $tempdir/$file
+            else
+                mv $tempdir/$file .
+            fi
             ;;
     esac
 done
